fix bug in OpenCV rotation
The center of the face was being reported as the center in the internally-cropped image, not the center
in the original mask. When the code that detects which face corresponds to which mask ran, it samples
the mask at the center of each face, but since the center was wrong, it would use the default index 0.